+++ core/src/main/java/arez/component/MemoizeCache.java
@@ -241,7 +241,6 @@ public final class MemoizeCache<T>
                                   null,
                                   onDeactivate,
                                   null,
-                                  null,
                                   _priority,
                                   false,
                                   true,