+++ integration-tests/src/test/java/arez/integration/computed_value/ComputedNonArezDependenciesTest.java
@@ -34,7 +34,6 @@ public class ComputedNonArezDependenciesTest
                         null,
                         null,
                         null,
-                        null,
                         Priority.NORMAL,
                         true,
                         true,