+++ integration-tests/src/test/java/arez/integration/computed_value/ComputedValueKeepAliveTest.java
@@ -35,7 +35,6 @@ public class ComputedValueKeepAliveTest
                         null,
                         null,
                         null,
-                        null,
                         Priority.NORMAL,
                         true,
                         true );