+++ CHANGELOG.md
@@ -7,6 +7,9 @@
   in source code analysis programs such as spotbugs.
 * **\[gwt-output-qa\]** Add assertion to verify that the `arez.*Info` interfaces are not present if
   `Arez.areSpiesEnabled()` returns false.
+* **\[core\]** Remove the `onDispose` parameter from the `ArezContext.autorun(...)` methods and the
+  `ArezContext.computed(...)` methods. This hook method was never used from within the component model
+  and infrequently used (if ever) from the raw API. Removing it reduced code complexity and size.
 
 ### [v0.106](https://github.com/arez/arez/tree/v0.106) (2018-08-31)
 [Full Changelog](https://github.com/arez/arez/compare/v0.105...v0.106)